Use fixed-sizes integer types

+++ src-core/common/dsp/file_sink.h
@@ -68,12 +68,12 @@ namespace dsp
             if (d_sample_format == WAV_16)
             {
                 // write wav header
-                int subchunk1_size = 16;
-                int audio_format = 1;
-                int num_channels = 2;
-                int bits_per_sample = 16;
-                int byte_rate = samplerate * num_channels * bits_per_sample / 8;
-                int block_align = num_channels * bits_per_sample / 8;
+                uint8_t subchunk1_size = 16;
+                uint8_t audio_format = 1;
+                uint8_t num_channels = 2;
+                uint8_t bits_per_sample = 16;
+                uint64_t byte_rate = samplerate * num_channels * bits_per_sample / 8;
+                uint8_t block_align = num_channels * bits_per_sample / 8;
 
                 output_file.write("RIFF----WAVE", 12);
                 output_file.write("fmt ", 4);
@@ -120,8 +120,8 @@ namespace dsp
             if (d_sample_format == WAV_16)
             {
                 // complete wav header
-                int data_size = get_written();
-                int chunk_size = data_size + 36;
+                uint64_t data_size = get_written();
+                uint64_t chunk_size = data_size + 36;
 
                 output_file.seekp(4);
                 output_file.write((char*)&chunk_size, 4);
